Seperating order tests
